+++ src/Places/ServerVariable.php
@@ -75,16 +75,40 @@ class ServerVariable extends Place_Base {
 
 		// bail if no name for the server variable is given.
 		if ( empty( $config['server_variable'] ) ) {
+			// log this error.
+			$this->get_crypt_obj()->add_error(
+				'server_variable_missing',
+				'The server variable is missing in configuration for Crypt for WordPress.'
+			);
+
+			// do nothing more.
 			return;
 		}
 
 		// bail if given value is not a string.
 		if ( ! is_string( $config['server_variable'] ) ) {
+			// log this error.
+			$this->get_crypt_obj()->add_error(
+				'server_variable_not_a_string',
+				'The server variable is not a string.'
+			);
+
+			// do nothing more.
 			return;
 		}
 
 		// bail if given server variable does not exist.
 		if ( empty( $_SERVER[ $config['server_variable'] ] ) ) {
+			// log this error.
+			$this->get_crypt_obj()->add_error(
+				'server_variable_missing_in_server',
+				'The variable is missing in server variable. Did you miss the configuration in your hosting?',
+				array(
+					'server_variable' => $config['server_variable'],
+				)
+			);
+
+			// do nothing more.
 			return;
 		}